pyproject.toml
[tool.poetry]
name = "wapdc"
version = "0.1.0"
description = ""
authors = ["GGOMESSANTOS <160145552+GGOMESSANTOS@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"# Ou a versão mais recente disponível
office365-rest-python-client = "^2.3.11" 
pymongo = "^3.12.0"  # Ou a versão mais recente
mysql-connector-python = "^8.0.28"  # Ou a versão mais recente
psycopg2 = { version = "^2.9.1", extras = ["binary"] }  
ruamel-yaml = "^0.18.6"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
pyarrow = "^18.1.0"
xlsxwriter = "^3.2.0"
boto3 = "^1.35.81"
azure-storage-blob = "^12.24.0"
pyspark = "^3.5.3"
delta-spark = "^3.2.1"
iceberg = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
